fix: Checksum errors with some MCP2-generated card images (#84)
* test: Add test case

* refactor: Rename XOR hash package

* fix: Change how checksums are generated on read

Checksums were previously generated after reading. The problem with this
was the blank fields can be non-zero in some cases with images generated
by the MemCardPro2 but they would be assumed to be zero when generating
the checksum which would then cause it to not match the checksum
embedded in the image.

Instead generate the checksum whilst reading using a copy of the actual
bytes.